Use node.jumps(), not node.isStatement() to determine if statement is pure.
For loops and switches are always considered statements, so we missed
lint errors where a function ends in a for loop or a switch.

From a quick look over nodes.coffee, it seems jump() is truthy for
return, continue, break (in the cases that they result in a pure statement),
and all other nodes jump() if they have a child node that jump()s, so
this seems to be identical to the pureness of the expression.

Now would be a good time to add some tests.

Closes #4, closes #5